move compatible test
